Uploading Submission

+++ Submission/Classes/RandomNoGen.cs
@@ -10,17 +10,17 @@ namespace BVC_Filters
     {
         private static Random RandomGenerator { get; set; }
 
-        public static UInt64 Int64Gen()
+        public static int Int64Gen()
         {
             if (RandomGenerator == null)
                 RandomGenerator = new Random();
-            var buffer = new byte[sizeof(UInt64)];
+            var buffer = new byte[sizeof(int)];
             RandomGenerator.NextBytes(buffer);
-            return BitConverter.ToUInt64(buffer, 0);
+            return BitConverter.ToInt32(buffer, 0);
         }
 
-        public static List<UInt64> RandomList(long count) {
-            List<UInt64> list = new List<UInt64>();
+        public static List<int> RandomList(long count) {
+            List<int> list = new List<int>();
             while (count-- > 0)
             {
                 list.Add(Int64Gen());
